Main functionality of intruder detection
This lambda is triggered when a detected image is added to S3. When this happens tyhe whitelist images are compared to this detected image to see if it is an owner or intruder. If it is NOT the owner and/or the secutiry level allows for notifications to be sent out, then it sends a notificaiton based on whether the user wants an SMS, email or push notification.

From the previous update, the following functionality has been added:
1. get user data such as whitelist images, preferences and security settings
2. send email and sms
3. retrieving the metadata from the S3 object that triggered the function.
4. add the detected image to the database.
5. append logs about the event that occured
6. remove object from s3 if it is the detected image is an owner.

TODO:
Add try blocks for corner cases.
Update the DB format for notificaitons to allow only Cognito values.