Update Location endpoint function
This lambda function is triggered when a location in the HCP is updated. A location in our context is a location where a cluster of cameras may reside in a user's home where the HCP is running.

In order to achieve this, the current control panel is retrieved with the current locations (and cameras) and is checked to see if the new location already exists. If it does not, then the old location is replaced with the new location with the respective cameras from the old location.

This will essentially update the location in which a camera captures videos.