Remove Camera from given site and location
The lambda handles the deletion from the UserData table. The endpoint requires the camera id, site_id and location to be passed as parameters. This is used to uniquelly delete a camera in the dynamo table at the given location and site using the update_item, with the unique update expression 'REMOVE' to only remove a given column.