Optimisations & Watchlist Detection Feature
This updated commit includes optimisations that significantly decreaased the latency, and inclustion of the watchlist of specific users.

The optimisation was improved by removing the request to get the training locations from the API, to having it in the function itself. This allows for the data to be reterieved directly from dynamo instead of having to go through the gateway on the detectIntruer endpoint and invoke another lambda to reterieve the information prodided.

The additonal feature added is only applicable to when the security level is 1. This means that the intruder detection module only checks to see whether the detected face is an owner or not, and if it is not it sends off a notificaiton that there is a possible breach about to occur. The new feature extends this idea such that when a face is detected to be a known whitelist image, it checks to see whether the whitelist image is on the watchlist. If it is the case that this is true, it sends off a notification with a custom message that the user has provided in the frontend.