Detected Image preprocessing & updated log structure
The HCP does some preprocessing to ensure that images aren't sent in bursts, however, it does not restrict this entirely. Thus, we have included an extra preprocessing step to ensure that detected images within the past 10 minutes aren't the same as the current detected image being sent through intruder detection. This is done by getting the detected images in the past 10 minutes, and comparing that with the detected image to see any of them have faces that match the detected face. If so, then it does not store the detected image, else, i runs the normal intruder detection as before.

The logs have also been updated. This is due to the fact that we are not tagging detected images through the logs anymore, but rather have a separate page in the Web app. Hence, the logs do not need extra information such as the camera_id etc.