Watchlist Functionality
This is the functionality of adding whitelist images to the watchlist. This means that when a detected face is found in the HCP, it sends this image to the file system and this triggers the intruder detection module. If it is the case that the detected image is an owner then it checks to see if it is in the watchlist. If this is so, it sends out a customised message to the user based on what they have specified on the frontend.

Hence, this update allows for the insertion and toggling of the watch boolean variable in the database for the intruder detection analysis to check this.