Detected images GET for Tagging of detected images as whitelist
This commit allows for the GET of the detected images to be tagged in the frontend. Before we were doing this through the logs, however, this may have caused update anomolies as the logs rely on the detected images and the tagging of them would rely on the logs, which inherently rely on the detected images.

Specifically, this route returns the key, presigned url, location of where the image was captured, and the timestamp it was captured. To tag the image as being a known whitelist in the frontend, the same route will be used, but instead of GET, it will be POST, and will trigger the DyanmoPost lambda. The POST requires the key name, and a unique name of the tagged person. This is returned by this GET method. The timestamp and location and merely for filtering purposes and the URL is for displaying the image in the frontend.