Restructuring Identities for owner analysis
This commit involves the updated structure for the updated structure of the identities. This structure allows us to be able to run analytics for the the profiles of the system, such that each time an owner is identified, it stores it in it's own list so that the frontend can view this as time-series data on a line graph.